Introduce new CodeMirror block for editing code.
While demonstrating Gutenberg to my team the question was asked about
editing code blocks: "it's so terribly difficult in the WordPress
editor."

Wanting to demonstrate some of the flexibility that the block system
provides I created a new block type which turns a normal block into an
instance of the CodeMirror editor.

This should eventually replace the standard `code` block once it's ready
for production. At the moment it lazy-loads the `CodeMirror` editor
_and_ the mode files, so thankfully this keeps the initial builds small.